requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
cachetools==6.2.3
certifi==2025.11.12
charset-normalizer==3.4.4
distro==1.9.0
flask==3.1.0
flask-socketio==5.4.1
flask-cors==5.0.0
pillow==11.0.0
pillow-heif==0.20.0
frozenlist==1.8.0
google-auth==2.43.0
google-genai==1.55.0
gradient==3.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.0
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
websockets==15.0.1
yarl==1.22.0
